requirements.txt
PyMySQL
PyYAML
requests
telepot
paramiko
pytz
APScheduler
scandir